+++ packages/driver/src/cy/commands/querying/root.ts
@@ -1,5 +1,6 @@
 export default (Commands, Cypress, cy, state) => {
-  Commands.addQuery('root', function root (options: Partial<Cypress.Loggable & Cypress.Timeoutable> = {}) {
+  Commands.addQuery('root', function root (options: Cypress.LogTimeoutOptions) {
+    options = options || {}
     const log = Cypress.log({
       timeout: options.timeout,
       hidden: options.log === false,