+++ packages/errors/test/unit/visualSnapshotErrors_spec.ts
@@ -1112,6 +1112,11 @@ describe('visual error templates', () => {
         default: [1, 'chrome', 62],
       }
     },
+    CDP_FIREFOX_DEPRECATED: () => {
+      return {
+        default: [],
+      }
+    },
     BROWSER_PROCESS_CLOSED_UNEXPECTEDLY: () => {
       return {
         default: ['chrome'],